+++ car_physic/__init__.py
@@ -1,45 +1,17 @@
-import dataclasses
 import math
 import typing
 
 import numpy as np
 
-
-@dataclasses.dataclass
-class Motor:
-    forward_acceleration_factor: float
-    back_acceleration_factor: float
-    max_acceleration: float
-    max_velocity: float
-    acceleration: float = 0.0
+import car_physic.piece.brakes
+import car_physic.piece.chassis
+import car_physic.piece.motor
+import car_physic.piece.steering
 
 
-@dataclasses.dataclass
-class Chassis:
-    length: int | float
-    size: np.ndarray
-    position: np.ndarray
-    mass: float
-    angle: float = 0.0
-
-
-@dataclasses.dataclass
-class Brakes:
-    brake_deceleration: float
-    brake_hand_deceleration: float
-    free_deceleration: float
-
-
-@dataclasses.dataclass
-class Steering:
-    max_steering: int
-    performance: float = 30
-    direction: float = 0
-
-
-DEFAULT_GENERIC_MOTOR = Motor(1 / 1000_000_00, 2 / 1000_000_000, 1 / 5, 1 / 100)
-DEFAULT_GENERIC_BRAKES = Brakes(1 / 100000, 1 / 50000, 1 / 200000)
-DEFAULT_GENERIC_STEERING = Steering(120, 80)
+DEFAULT_GENERIC_MOTOR = car_physic.piece.motor.Motor(1 / 1000_000_00, 2 / 1000_000_000, 1 / 5, 1 / 100)
+DEFAULT_GENERIC_BRAKES = car_physic.piece.brakes.Brakes(1 / 100000, 1 / 50000, 1 / 200000)
+DEFAULT_GENERIC_STEERING = car_physic.piece.steering.Steering(120, 80)
 
 
 def rotate_vector(vector: np.ndarray | list | tuple[int, int], radians: int | float) -> np.ndarray:
@@ -50,16 +22,16 @@ def rotate_vector(vector: np.ndarray | list | tuple[int, int], radians: int | fl
 
 
 class CarPhysic:
-    def __init__(self, chassis: Chassis, motor=DEFAULT_GENERIC_MOTOR, brakes=DEFAULT_GENERIC_BRAKES,
+    def __init__(self, chassis: car_physic.piece.chassis.Chassis, motor=DEFAULT_GENERIC_MOTOR, brakes=DEFAULT_GENERIC_BRAKES,
                  steering=DEFAULT_GENERIC_STEERING):
         self.velocity: np.ndarray = np.array([0, 0], dtype=np.float32)
         self.force: np.ndarray = np.array([0, 0], dtype=np.float32)
         self.angular_velocity = 0
 
-        self.chassis: Chassis = chassis
-        self.motor: Motor = motor
-        self.brakes: Brakes = brakes
-        self.steering: Steering = steering
+        self.chassis: car_physic.piece.chassis.Chassis = chassis
+        self.motor: car_physic.piece.motor.Motor = motor
+        self.brakes: car_physic.piece.brakes.Brakes = brakes
+        self.steering: car_physic.piece.steering.Steering = steering
 
     def process(self, dt: float, throttle: int | float, brake_hand: bool, steering: int | float) -> typing.Tuple[
         np.ndarray, float]:
