+++ car_physic/piece/chassis.py
@@ -1,0 +1,10 @@
+import dataclasses
+import numpy as np
+
+@dataclasses.dataclass
+class Chassis:
+    length: int | float
+    size: np.ndarray
+    position: np.ndarray
+    mass: float
+    angle: float = 0.0
