+++ car_physic/piece/steering.py
@@ -1,0 +1,8 @@
+import dataclasses
+
+
+@dataclasses.dataclass
+class Steering:
+    max_steering: int
+    performance: float = 30
+    direction: float = 0
